pyproject.toml
[project]
name = "aivtuber-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.194.0",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
]
